+++ srcs/main/vector_utils2.c
@@ -6,7 +6,7 @@
 /*   By: qsomarri <qsomarri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/27 16:19:32 by qsomarri          #+#    #+#             */
-/*   Updated: 2025/03/27 16:29:26 by qsomarri         ###   ########.fr       */
+/*   Updated: 2025/03/31 14:51:43 by qsomarri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 t_vector	normalize_vect(t_vector v)
 {
-	double	a;
+	float	a;
 
 	a = vect_norm(v);
 	return ((t_vector){v.x / a, v.y / a, v.z / a});
